(PUP-8696) Emit a deprecation warning for each function
Previously, we would only warn once the first time loadall, load or
loaded? was called for any function, which means as soon as you fix one
deprecation, you run into another one.

This commit ensure we emit at most one warning per autoloader method,
and for load/loaded? the name of the function. We could also include the
caller's name, but that seems very fragile as we can't guarantee that
the caller is a function.