(PUP-8780) Allow user type to work if sshkeys_core module is not installed
While the sshkeys_core module is going to be re-packaged into
puppet-agent, there is no guarantee that the type will be available if
Puppet is run from source. We now provide a warning if a feature of the
user type that depends on one of the types in the sshkeys_core module is
used while the type from the sshkeys_core is not available.