(PUP-8908) Include failed_to_restart status in Puppet::Resource::Status serialization
Previously, we were only including a count of resources that had the
failed_to_restart status in the metrics section of the report, but were
not marking the individual resource statuses with whether or not they had
failed_to_restart. This made it rather annoying to find which resource had
failed to restart.

Now we include failed_to_restart in the (de)serialization of a
Puppet::Resource::Status so that we can directly find out which
resource(s) contributed to the total failed_to_restart in the metrics.