(maint) Delete zone test
This test ensures that zone is correctly loading the zfs module, as that
is a dependency of zone. However, since we're extracting these modules,
zfs isn't available to zone like it used to be (as a part of the puppet
code base). Zone is in the process of being removed, which moves all
these tests to that module. This will include loading ZFS as a separate
module, preserving this functionality there. Rather than wait for the
zone removal, I'm just removing this test now.