(PUP-8090) Load starting environment into the default domain
Fixes an issue where 'puppet apply' starts in the 'cli' domain, compiles
the catalog - which switches to an environment-specific domain, then
deletes it - and resets to the default domain. Type/provider evaluation
after the compile then happens in the default domain, without any module
code.

Switch to loading the configured environment into the default domain.
That should affect various workflows as follows:
* 'puppet apply' starts in default domain and loads module translations,
switches to environment-specific domain and reloads the same module
translations for compilation, then switches back to the default domain.
This should ensure the entire run happens with all module translations
from the configured environment loaded.
* 'puppet agent' starts in default domain and loads module translations,
then switches to 'agent' domain and loads only vardir translations. This
ensures that only vardir translations are used, not any other installed
module translations. The default domain should not be used the rest of
the run.
* 'puppet master' starts in default domain and loads module code; during
each compilation it switches to a specific environment and uses
environment-specific translations, then switches back to the default
domain. The default domain will contain extra translations, but they
shouldn't be used for anything.
* Other puppet applications should start in the default domain with
module translations and stay in that domain the entire run.