(PUP-8748) Configure connection verify mode when making a request
This commit makes the SSL configuration of the Puppet::Rest::Client
configurable by the caller. By default, the client is not configured to
use any certs but has VERIFY_PEER enabled, which will result in an SSL
Error if a caller tries to make a request without first properly
configuring the client. If the client is configured to use VERIFY_NONE,
the state will reset to VERIFY_PEER after the request is made.
This ensures that we will not make insecure requests by mistake.