(PUP-8748) Create a new rest client for each SSL bootstrap phase
This commit ensures that we get the correct SSL configuration for each
request made during the SSL bootstrapping process. When no certs exist
on the host, VERIFY_NONE is used. From then on VERIFY_PEER is used,
first with no CRL configured, then with a CRL once that has been
downloaded.