(PUP-8942) Support issuing certificates with IP SAN
Currently, there is support for signing CSRs with a DNS subject alternative name, but there is no option to generate a cert with an IP SAN.
We are also currently using the -dns-alt-names flag, rather than using -subject-alt-names and having the user specify the altname.
The --allow-dns-alt-names flag is also misnamed.

This ticket adds IP SAN functionality, and aliases -dns-alt-names to -subject-alt-names. Users who do not specify their altname type will default to DNS, in order to support the old format.
If both are specified, -dns-alt-names will be ignored. --allow-dns-alt-names has been aliased to --allow-subject-dns-alt-names. Deprecation warnings were added to both flags explaining this change.
The specs were also modified to reflect this change.