(PUP-8950) Load rest client as soon as possible
Previously, we were lazily loading the rest client to avoid LoadError
issues when running under Puppet Server.

However the rest client monkey patches OpenSSL and when lazily loaded
will cause errors if the X509::Store is instatiated before it is monkey
patched.

This updates us to load the client prior to initializing the Store.