(PUP-8952) Respect server_list setting when using Rest::Client
This commit updates the server resolution logic of the
Puppet::Rest::Client to respect the `server_list` setting. Previously
is was falling back to the `server` setting (or whatever server it was
configured with) even when `server_list` was set and supposed to take
precedence.

This commit also updates the acceptance test to verify that
`server_list` is preferred first, to avoid hiding that large bug behind
less consequential failures.