(PUP-8939) Ensure Administrators is owner of Puppet Settings as root
Previously, when running as an Administrator it was possible to create settings
files and directories that were only owned by a single Administrator.  This may
make sense for normal file resources, it doesn't really apply to Puppet settings
files. Therefore if we are running with elevated rights (Puppet.features.root?)
and on Windows (Puppet.features.microsoft_windows?) then all settings files and
directories should be owned by the Administrators group, with the primary group
of LOCAL SYSTEM.  This is the typical configuration for files created as service
logged on as LOCAL SYSTEM. Note that the LOCAL SYSTEM user is always an implicit
member of the Administrators group therefore they don't needed to be added to
the resultant ACL.  Note we also use the SID representation to avoid
localization issues or account renaming.