(PUP-8942) Provide support for IP subject alt names
Currently, CSRs can only be signed using DNS type subject alt names (referred to as dns_alt_names or dns-alt-names).
This commit provides support for IP SANs, but does not address any of the naming issues brought up in this ticket.
More information on the blocking factors for that name change can be found at #6888.