(docs) edit Puppet license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit makes a couple changes so that Licensee recognizes Puppet's
license type as Apache 2.0. It updates LICENSE so that it stores only
the text of the Apache license in its entirety. It also updates the
README to hold some of the info that was previously stored in LICENSE
and that does not exist elsewhere. The updated README section has been
renamed to "Licensing", so that Licensee correctly recognizes LICENSE
as the only license-related file.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>