(PUP-8971) Add Encrypted data type and decrypt function
This adds an Encrypted data type and a function decrypt() that
can decrypt such a value.

An Encrypted is only equal to itself.

It accepts an options hash with `cipher` and `node_name` as optional
keys; `node_name` is the catalog requesting node when in master
mode and the local node name when doing an apply.

The `cipher` is 'AES-256-CBC' by default.

The `decrypt` function can decrypt an `Encrypted`. The decryption is for
the localhost unless a node name is also given (a private key must then
be available for that node name).

The typical use case is that a decrypted value needs to be handled such
as the secret is not revealed by accident. Therefore, the returned value
from `decrypt` is always a `Sensitive` value as that prevents it
from being accidentally revealed in reports and logs.

An `Encrypted` is `RichData`.

The test for `Encrypted` includes testing the `decrypt` function since
all of the functionality (except being a function) is in the data type.

The test for `Encrypted` includes testing round-trip in catalog since it
requires the same set up of CA and having signed certificates as the
rest of the tests for `Encrypted`.

This also modifies `indirected_routes.rb` and `trusted_information.rb`
such as the certificate from a remote request is made available to
`Encrypted`. The altered logic also handles `puppet apply` by using the
certificate for the `localhost`.