(PUP-8976) Make it safe to use __ptype as a key in a hash
Before this, if a Hash had a key `__ptype` it would be taken as a
special type marker. That was bad because it makes some keys reserved in
a hash.

In this commit, such hashes are treated as having "non string keys"
which means it uses an alternative representation. This avoids the
problem when the rare cases there is such a hash occurs.