(PUP-8974) Safely deserialize stringified array
The augeas provider used Kernel#eval to convert stringified arrays to
Ruby arrays. For example, it extracted the array part of the "clause"
below:

    onlyif => 'values HostKey == ["/etc/ssh/ssh_host_rsa_key"]'

and called Kernel#eval with '["/etc/ssh/ssh_host_rsa_key"]'. Using eval
is bad because it executes arbitrary code.

This commit changes the provider to convert the comma delimited
string to a Ruby array. Note it will accept unbalanced quotes, so input
like "[\"a', \"b']" is converted to ['a', 'b']