(PUP-8985) Default manage_internal_file_permissions to false on Windows
Previously the MSI packaging for Puppet Agent changed the default for the
manage_internal_file_permissions setting to false as part of CVE-2018-6513.
However puppet still had a default of true.  This commit changes the default to
false for Windows platforms, to mirror that of the packaging, and to not undo
the remediation for the CVE.