(PUP-8988) Don't pluginsync vendored modules
Previously, vendored modules behaved like other modules in the
modulepath. When acting as a server, agents would pluginsync the
vendored modules. This can cause problems if the vendored module is not
compatible with an older agent.

This commit makes vendored modules behave as though they are part of
puppet. They are not pluginsynced, but the autoloader can resolve
types, providers, etc from vendored modules. Vendored modules always
have least precedence.