(PUP-8971) Add support for Error.cause
In Ruby >= 2.1.0 there is support for `cause` on all raised errors.
This is very useful for keeping track of the original exception.

This should really be cleaned up in the puppet code base when no longer
supporting Ruby < 2.1.0.

This adds support for `cause` in one of the error classes to make it
available for one of the tests that were failing on older rubies.