(PUP-9010) Ensure block is always run by with_text_domain
`with_text_domain` is used to compile a catalog. In the refactor to use
`with_text_domain`, compiling a catalog when `disable_i18n=true` was
broken because `with_text_domain` would not evaluate the block passed to
it. Fix so it always evaluates the block, i.e. compiles a catalog.

Update acceptance test to ensure server works with `disable_i18n=true`.