(PUP-8780) Remove code related to extracted sshkeys_core module
This also removes unit & integration RSpec tests from the User type, as
they have been moved into the sshkeys_core module, and would require
having the module installed & available to the test suite to be able to
run them as part of the normal Puppet tests. The setup required is not
something that is normally done, so those tests would basically never be
run here.