(PUP-8971) Simplify error handling for unsupported cipher
This modifies the error handling such that the error message
is the same for a cipher that cannot be used irrespective of
if it is not available at all, or if not acceptable. It also shows
a list of both available and acceptable ciphers.

Before this it may first give an error for unavailable, only to then
possibly error saying that the cipher is not acceptable. Now all
checking is done on the union of acceptable and available ciphers.