(PUP-8971) Add feature to select best cipher when encrypting
It is now possible to give Encrypted an array of "target available
ciphers" as the `cipher` option, from which Encrypted can choose. It 
will select the most preferred according to the `accepted_ciphers`
setting, and if that setting is an empty array, the first given cipher.

The "target available ciphers" is for the user to compute. Either via
having a fact that produces them, or as data in hiera.