(PUP-9051) Fix SMF provider to properly implement enableable semantics
Previously, the SMF provider had the same semantics for enable as it did
for ensure. Specifically, #enable and #disable delegated to #start and #stop.
This is completely inconsistent with the enableable semantics for service
providers, semantics which other Unix-based service providers like systemd
implement. For example, if you had a SMF-based service resource with
enable => true and ensure => false, the net effect of the operation would
be a disabled service that would remain disabled when rebooted, which is
incorrect and inconsistent. The correct and consistent effect should be
a disabled service that would start up when rebooted.

Essentially, enableable semantics requires that enable and ensure be
independent operations where enable handles whether the service starts
or stops at boot time, while ensure handles whether the service starts
or stops in the current running instance. The previous SMF provider was
conflating the two properties together so that enable and ensure were
no longer independent operations.

This commit fixes that. It rewrites core parts of the SMF provider to
now conform to enableable semantics and also fixes some possible issues
with it. The key takeaways here are:

* Service commands now use the service FMRI instead of the resource
name. This is the unique resource identifier for the particular
instance. Previously, we'd pass in the resource name instead which would
match a pattern corresponding to multiple service instances. We only
want to manage one service instance in a single service resource. The code now
fails if more than one FMRI exists for a given service resource name.

* We now sync. the enable and ensure properties of the service resource in a single
transaction, specifically inside the #flush method. This is due to the dependency
between the ensure and enable properties caused by how SMF works under the hood.

Acceptance tests for this new behavior have been added. The existing
nonexistent service tests on the AIX service provider have been
refactored to a common helper function that both the SMF and AIX
provider tests can call on. This helper function should provide a clean
interface for nonexistent service tests that may be used for other
service providers in the future.