(PUP-8988) Define vendormoduledir as a string
Calling `Puppet.settings.use(<section>)` will create files and directories
for all of the `:file` and `:directory` based settings in that section.
Previously, running `puppet lookup` as a non-root user failed, because
lookup uses the `master` section, the `vendormodulesdir` setting is
defined in the `master` section, and the user did not have permission to
create the `/opt/puppetlabs/puppet/vendor_modules` directory.

This commit changes the vendormoduledir setting to be of type `:string`.
This way puppet never tries to manage its existence via a settings
catalog. It doesn't use `:path`, because that settings type is reserved
for settings that allow multiple directories separated by
File::PATH_SEPARATOR.