(maint) Verify autoload returns module unique to one path
Previously the test verified that a valid module in multiple module path
directories are returned, while an invalid module is not. Expand test to
include the case where a module is only in one module path.

Also clarify that "dir_a", "dir_b", etc correspond to directories
containing a set of modules whose names are "one", "two", etc.