(PUP-9018) Return the first module in the modulepath
Previously, Autoload.module_directories could return duplicate modules
if a module was installed in multiple directories in the modulepath.
Worse if there were different versions of a module installed, then the
Autoloader could load a type from one version of the module, but the
provider from a different one. For example, consider v1 of a module
containing type foo, provider bar, and v2 of the same module containing
type foo, and providers bar and baz. If module v1 comes before v2 in the
modulepath, then puppet will load type foo from v1, but provider baz
from v2. Undefined behavior can result if the v2 provider expects v2 of
the type.

The autoloader behavior is also inconsistent with how the rest of puppet
works. For example, when creating ModuleLoaders we use
`environment.modules`, so only the first module in the modulepath is
visible[1]. Similarly, when serving files from modules, loading facts
from modules, installing modules via PMT and resolving module
dependencies.

This commit uses `Environment#each_plugin_directory` to enumerate the
lib directory for all modules in the environment. As a result of this
change, modules with invalid metadata or invalid module names will not
be included. Those without any metadata continue to be included.

Also the returned module directories will be normalized using
Autoload.cleanup, as is consistent with other usages of
`each_plugin_directory`.

[1] https://github.com/puppetlabs/puppet/blob/5.5.3/lib/puppet/pops/loaders.rb#L415-L420