(PUP-8973) Add a CatalogCompiler to PAL
This adds support to PAL such that it is possible to compile
a catalog by using `with_catalog_compiler` instead of
`with_script_compiler`.

This also adds the ability to inject `compiler_param` (a catalog
compiler) and `pal_compiler` (a catalog or script compiler)
into a 4.x function.

This is the first commit in a series of commits. This adds the
basic foundation and allows obtaining the catalog in "ral" form.