(PUP-8997) Deprecate the `cert` subcommand
This commit deprecates the `cert` application, which will be updated in
Puppet 6 to error with instructions on how to use the Puppet Server CA
CLI to do the actions that `cert` currently performs.

Unlike with face, calling the `deprecate` message in an application only
causes a warning to be issued when it is run, and does not update the
summary (and therefore not the help output), so this commit updates the
summary to indicate the deprecation.