(PUP-9008) Use `initctl version --quiet` to detect upstart daemon
Previously, we'd confine the upstart provider to platforms where the
/var/run/upstart-socket-bridge.pid file exists because that would mean
that the upstart daemon was running. This was not quite correct. The
upstart daemon could still be running even if that file did not exist
(e.g. such as on a RHEL 6 machine, for example).

We can use `initctl version --quiet` instead to check if the upstart
daemon's running. If that command fails, then the upstart daemon isn't
running so we cannot use the upstart provider. If it succeeds, then the
upstart daemon is running so we can use the upstart provider. We confine
our provider to platforms that result in a successful execution of this
command.