(maint) Add with_libdir helper
Puppet will modify $LOAD_PATH whenever Puppet[:libdir] is modified. Add a helper
method to restore $LOAD_PATH and the libdir setting when those tests complete.
Also modify tests to not be sensitive to the current $LOAD_PATH, as it changes
during the test run. Instead verify that the public `search_directories` method
starts with gem directories and the libdir, and ends with the vendored module
lib directories.