(PUP-9015) Prepend libdir and append vendored modules
Prepend Puppet[:libdir] and append all vendored modules 'lib' directories to the
ruby $LOAD_PATH. Note previously libdir was appended, though the order
depended on the order that the hooks were executed. We append vendored
modules to the LOAD_PATH so that requiring helper code within the same
module works as expected.