(maint) Fix jruby unit test failures caused by PUP-9008
PUP-9008 introduced a :confine block that shelled out to initctl
to check for the existence of the upstart daemon in order to see
if it is a suitable provider for the current platform. This called
Puppet::Util::Execution.execute which calls Kernel.fork. Kernel.fork
is not supported in jruby.

Many unit tests ended up loading the upstart provider's .rb file as part
of loading the other service providers. Thus, the :confine block added
in PUP-9008 was evaluated, and Kernel.fork was called. These tests
would fail on jruby, which is one of the ruby versions CI tests on.

This commit fixes these unit tests. It applies the following changes.
* All service provider tests are skipped if we're running the unit
tests on jruby. Many of these tests load the provider in the describe
block, which means we cannot use rspec's :if or :unless options because
Ruby would evaluate the load expression before getting to those options,
causing the test to immediately fail on jruby due to our :confine block.
Thus, we modify these tests to move the loaded provider into a
`provider_class` variable and use that instead of `described_class`. The
test description now uses a string instead.

* Any other tests that happen to use a service provider stub out
the execute method.