(PUP-9642) explicitly set the confdir permissions
The `puppet device` application is using the `confdir` setting to have
puppet create a new private directory for the device's use. This causes
the directory to be created for each new device through the magic of
`Puppet.settings.use`.

* The defaults do not specify a filemode for `confdir`.
  Therefore any environmental markers, like `umask`, get propagated
  through to the file system.
* The settings subsystem switches to the configured `user`/`group` before
  writing files, and PE is setting those configuration values to non-root
  values.
  Therefore any filesystem operation has to pass through filesystem
  permissions.

With `umask 0027` the `$devicedir/$devicename` directory gets created as
`root/root` with `drwxr-x---` permissions, and the subsequent attempt to
write the private key into a subdirectory fails.

This commit adds an explicit chown for the newly created confdir that
allows access, thereby allowing the subsequent SSL key operations to
succeed.

Since the SSL keys are independently protected in their directories,
this change has no effect on the overall security.