(PUP-8977) Handle empty apply statement
Previously, ApplyBlockExpression was inheriting some behavior from
BlockExpression while not necessarily conforming to its expectations.
Specifically, it has a "statements" attribute, but it was possible for
that attribute to be nil, whereas a BlockExpression requires it to
always be an array. We now explicitly handle the empty apply block case
and use an empty array as the body.

This also removes some duplicated code that is inherited from the parent
class.