(PUP-9039) Properly construct apply statement body
Previously, we were using 'infer' to convert what was already valid as
an AST object into an AST object. That caused the body of an apply block
to always be interpreted as a LiteralList. Because of that, non-rvalue
expressions weren't allowed inside the apply block.

This commit also fixes the pcore representation of ApplyBlockExpression
to properly inherit from BlockExpression.