(PUP-9039) Allow virtual resource collection in apply blocks
Because virtual resource declaration and realization happens exclusively
in the scope of a single catalog, it's perfectly valid to use in an
apply block. This is especially important because of the amount of
existing module content that relies on virtual resources.

We now allow collection expressions in the task checker, so long as they
are within an apply block and collecting virtual resources. Exprted
resource collection is still disallowed within apply blocks.