(PUP-8977) Don't treat apply blocks as idempotent
Previously, if every statement in an apply block was idempotent
(for instance, one including only resource declarations), compilation
would fail if any other statement followed the apply block. This
required adding unnecessary variable assignments just to get the code to
compile.

In reality, an apply block is by definition *never* idempotent (it
exists purely for side effects) and it shouldn't be treated as such.
This change allows multiple apply blocks to be used in a row, without
adding intermediate variables.