(PUP-8528) Remove write only yaml node terminus
The write_only_yaml node terminus was used to "determine the list of
nodes that the master knows about"[1] and predated widespread puppetdb
adoption. The write_only_yaml has been deprecated[2] since 4.10.5, and
this commit removes it.

Note this should result in a puppetserver speedup as it no longer will
need to serialize node data as YAML to disk.

[1] https://projects.puppetlabs.com/issues/16753
[2] https://tickets.puppetlabs.com/browse/PUP-6060