(PUP-8683) Change default http read and run timeouts
Previously puppet could hang reading from a socket if the response never
arrived, for example if the TCP FIN or RST never arrived. Set a default
http read timeout to avoid that. Note the http connect timeout already
defaults to 2 minutes.

Similarly, set the watchdog timeout (runtimeout) to 1 hour. If a puppet
run takes more than that it will be killed by the watchdog. Overrun can
occur but at least puppet won't hang indefinitely, such as if puppet
executes a command that waits indefinitely. Prime example is a Windows
MSI that displays a UI waiting for user input that never happens.