(PUP-9035) Refactor specs for consistent before/after usage
Before this change, tests often utilized the `before(:all`) method
or the `around(:each)` method. This had the effect of running code
prior to any `before` blocks specified in the spec_helper. This
change normalizes tests to use consistent `before` and `after` methods
so the `before` block in the spec_helper consistently runs first.