(maint) Don't overly stub base yaml terminus
The TestSubject class is not an indirected type and doesn't have any
relationship to the YAML terminus we're testing. Use
Puppet::YamlTestModel along with its indirection and terminus instead.

Don't stub request objects, as the indirection name `:my_yaml` has no
relationship to the indirection name :yaml_test_model. Use `indirection.request`
as a shortcut to generating Request objects for that type of indirection.

Create yaml files so that we're not stubbing the code we're trying to test.