(PUP-7834) Safely load yaml_data function output
As described in PUP-7594 and commit 6dd67c3bd7622aae65da22cf093fafadb755bdc4
hiera hash keys can be symbols and are converted to strings. The
yaml_data function is used to load the data, and must accept symbols.

Previously, we always provided location information if YAML.load
produced a class the caller didn't expect. Now we only provide location
information if the unexpected class is a symbol, string, numeric, etc.
If an unsafe class would be produced, then we treat it is as a syntax
error and no location information is included.