(PUP-7834) Safely load pseudo-network yaml format
Ideally we could just remove the `yaml` network format, since we don't
allow YAML over the network. However, the formats are used in
non-networked contexts, such as faces. For example the following will
load facts from yaml and render them to the console:

    puppet facts find <name> --facts_terminus yaml

This commit safely loads YAML and restricts the allowed classes to
a subset of indirected types (catalog, report, node, facts, resource).

This commit preserves the behavior of raising a FormatError if the
intern or intern_multiple is invalid YAML, either due to syntax or if
the class is not allowed.

The list of accepted classes can't be a constant because the yaml
network format is evaluated before the other classes exist. The list of
accepted classes was generated using the following:

    Puppet::Indirector::Indirection.instances.map do |name|
      Puppet::Indirector::Indirection.model(name)
    end.select do |model|
      model.supported_formats.include?(:yaml)
    end - [Puppet::SSL::Host, Puppet::DataBinding, Puppet::Status]

This is based on a patch from Thomas Hallgren <thomas@tada.se>.