(PUP-9049) Refactor Puppet::Provider::AixObject class
This class is the base class for the AIX user and group providers.
Unfortunately, it is quite old, not written in an idiomatic style, and
is more complicated than it needs to be. Some examples include:

* The mapping for <puppet_property> => <aix_attribute> and vice-versa is
too complicated. The translation function is special cased to four cases
with poorly named variables.

* There are two different functions that parse out an AIX object. One is
from a colon-separated list of attributes, and one is from a
space-separated list of key-value pairs. We should only have one parsing
function, and that is the colon one. However, the colon parser is buggy.
It does not handle escaped colons (where they're escaped by #!). See
PUP-7241.

* The provider will work fine if a puppet property is specified as part
of the attributes property . For example, for a User resource if you
specify the uid property and then the id attribute in the attributes
property (where id <=> uid), Puppet will update the uid of the user
twice -- one with what you passed in for the UID property, and one with
what you passed in for the id property. This is undefined and incorrect
behavior, and was the issue of PUP-7533. We want the provider to error
if the user specifies a Puppet property in its AIX attribute form.

* The attributes property is difficult to reason about. The code
filter_attributes does two different things for a User resource and a
Group resource. The User resource references the original `should` parameter of
this property to filter out the attributes, while the Group resource
filters out the attributes that aren't managed by Puppet. This is
strange. Attributes have the same meaning for users and groups on AIX,
i.e. some are Puppet properties, some aren't. All of the attributes
related code should be contained in the AixObject class.

* There's line-by-line identical code copy-pasted in the AIX User and
Group providers. This should have been moved over to the AixObject
class.

etc. etc. etc. At a high-level, the AixObject class should handle all
attribute management logic, while the subclasses should handle any
custom properties they have that don't correspond to AIX attribtues
(e.g. like the password property for the User resource). We should also
define a strict separation between Puppet properties that map to AIX
attributes and AIX attributes that don't map to Puppet properties, where
the latter is stored as part of the attributes property.

This commit refactors the AixObject class to remedy all of these issues.
Specifically,

* The mapping between a Puppet property and AIX attribute is
simplified and declared in a DSL style (see the spec tests for an
example).

* We parse our attributes from the shell using the colon-seperated list
parser since we can have space-separated attributes. This parser fixes
PUP-7241

* Puppet now errors if a Puppet property is specified in the attributes
property.

* The attributes property is easy to reason about and all of the logic
that handles it is now in one place. The attributes property corresponds
to those AIX attributes that aren't mapped to Puppet properties.

* Much of the duplicate code shared between the AIX User and Group
providers has been moved to this class.

The unit tests have been rewritten and are a lot stronger than what they
were before.