(PUP-3021) Fix the KeyValue prop. to accept a Hash + AIX improvements
The KeyValue property class is a very old property class that likely
predated the Hash type in Puppet. It used to accept either a single
key-value pair (String) or an array of key-value pairs (array of
Strings). However internally, it would munge and manipulate these values
as hashes. This meant that when running e.g. puppet resource on a
resource that uses the KeyValue property, Puppet would output a hashed
param. for this property because that is what the internal retrieve
method returns. When this manifest was then compiled, Puppet would error
indicating that the property needs to be separated by the separator,
i.e. that Hash values are not accepted.

We should modify the KeyValue to accept a Hash value too, while also
preserving backwards compatibility. This commit does that. It moves over
the validation code to the main KeyValue property class and has it
accept only Hashes or Strings. We then munge our Hash to only contain
symbolic keys and string values, which is what the class already does to
a param. that's passed as an array of key-value pairs.

This commit also adds the option for the KeyValue property class to log
only the keys that changed in a given Puppet run. This is necessary for
e.g. the AIX attributes property in the User and Group resources, where
we don't want to print every single attribute in our log as there are
a lot. The AIX User provider would previously munge the returned attributes
value to return only the ones corresponding to should, if provided, so
that the property itself would only capture the keys that may change in
our Puppet run. This is bad. That type of logic should be contained in
the property class itself. We have that logic trigger on a class
variable that the subclasses can declare in their initialization block,
which we do for both the User and Group resources.