(PUP-7433) Reorganize bundler groups
Previously we excluded "development extra packaging" during CI runs,
which included some unnecessary gems from the "test" group, such as
ronn, yard, memory_profiler. And it excluded msgpack, which is an
optional feature-based dependency, which means tests were being skipped.

This commit reorganizes groups into:

    features
    test
    development*
    packaging*
    documentation*

where (*) means the group is optional and will only be installed when
requested as `bundle install --with <name>`.

This commit preserves all of the gems and constraints, except for
puppetlabs_spec_helper, which was unused and removed.

It also adds the jobs and retry options to .travis.yml, similar to what we
do for appveyor.