(PUP-7433) Move yard and rdoc out of documentation
The `doc` application can optional use `rdoc` based on a feature, so
move it to the features group.

The `doc:*` tasks emit a warning if it can't load the `yard` gem.
Resolving that is difficult because the tasks are defined using
`YARD::Rake::YardTask`. For now just move it to the `test` group. It's
also handy to run a yard server with --reload while making source
changes.

The `redcarpet` gem is no longer a dependency of anything we use, so
remove it.