(PUP-8320) Make filter less strict and accept truthy predicate
Before this, `filter()` was different from other functions taking
a predicate lambda in that it required the lambda to return
a boolean `true` to not filter.

This is now changed to make `filter` behave like the other functions
where any truthy value returned from the lambda means do not filter this
value.