(maint) Ignore X509 purpose during cert verification
The puppet cert, ca, etc subcommands have various methods for
determining if a certificate is valid. This is used to see which certs
have been signed, are still waiting to be signed, or have been revoked.
Historically, puppet set the purpose to PURPOSE_SSL_CLIENT on the
X509::Store used to verify certificates, but ruby+openssl 1.0 ignored
it. When using openssl 1.1, purpose is checked, which causes
CertificateAuthority#verify to raise an error when verifying the puppet
CA cert, since it doesn't have the 'SSL Client' purpose. This
corresponds to the 'id-kp-clientAuth' extended key usage extension[1].

However, puppet doesn't really need to care about the purpose when
determining if the cert has been signed, is expired, or revoked. This
commit changes the verify method to default to PURPOSE_ANY, and
preserves the behavior prior to openssl 1.1.

Note this does not affect TLS communication between the agent and
server. When acting as an SSL client, openssl always verifies that the
server's cert has the id-kp-serverAuth extended key usage extension. If
it doesn't, openssl sends a fatal TLS alert message of type 42 'Bad
Certificate' and closes the connection.

Cherry-pick f9e3b2ab4c1d6fe4f49c226de0f5185ec98a4d80 from PUP-9064

[1] https://tools.ietf.org/html/rfc5280#section-4.2.1.12