(maint) Deprecate settings related to master application
As part of PUP-8591 we removed the master application in Puppet 6. This
orphaned many settings that are not applicable to running a master with
Puppet Server or are better served living within its configuration
files.

To prepare for the removal of the puppet master application in Puppet 6,
this patch deprecates those settings.