(PUP-5985) Move windows gem initialization
Explicitly require windows specific gems in Puppet::Util::Windows
instead of relying on features.

This changes the semantics of the `microsoft_windows` feature from "are
we windows and could we load the prerequisite gems" to "are we windows".
However, the change is not meaningful, because puppet expresses its
windows-specific gem dependencies in .gemspec and Gemfile, when running
as a gem or from source, respectively. As a result, puppet couldn't
start if its prerequisites were missing.