(PUP-5985) Re-evaluate feature blocks on nil
Puppet has two ways of defining features. One version takes a list of
libraries:

    Puppet.features.add(:myfeature, libs: 'mylib')

The libraries are lazily required when `Puppet.features.myfeature?`
is first called. If it's successful, puppet will cache the result, so
future calls to `Puppet.features.myfeature?` are a noop. If it fails,
puppet will not cache the miss, and will try to require the libraries
the next time `Puppet.features.myfeature?` is called.

The other version takes a block and is evaluated when the feature is
added, and the results are always cached, even if the require fails:

    Puppet.features.add(:myfeature) do
      require 'mylib'
    end

If a feature is defined using `:libs` and a block, then the block is
used and the `:libs` are ignored.

This commit changes the following behaviors:

 * Feature tests defined using a block are delayed until
  `Puppet.features.<name>?` is called, as is done for `:libs`.

 * If a block raises a StandardError or ScriptError exception, then the
   result is not cached, just like is done for `:libs`. Note the `:lib`
   case could only raise LoadError which is a subclass of ScriptError.

 * If a block returns nil, then the result is not cached. True and false
   return values are always cached as they were before.

This is a change in behavior for any feature whose block returns nil or
raises a ScriptError or LoadError.

Also if `Puppet[:always_retry_plugins]` is false (as it is in
puppetserver), then the result of the feature test for blocks and
libraries are always cached.